Clean up utils tests.
