+++ tests/core/test_utils.py
@@ -80,82 +80,3 @@ def test_read_lines():
     lines = list(lines)
 
     assert len(lines) == 2
-
-
-os.environ["USER_NAME"] = "user"
-os.environ["PASS"] = "pass"
-
-
-def test_read_yaml_string():
-    config_without_env_var = """
-    user: user
-    password: pass
-    """
-    r = rasa.utils.io.read_yaml(config_without_env_var)
-    assert r["user"] == "user" and r["password"] == "pass"
-
-
-def test_read_yaml_string_with_env_var():
-    config_with_env_var = """
-    user: ${USER_NAME}
-    password: ${PASS}
-    """
-    r = rasa.utils.io.read_yaml(config_with_env_var)
-    assert r["user"] == "user" and r["password"] == "pass"
-
-
-def test_read_yaml_string_with_multiple_env_vars_per_line():
-    config_with_env_var = """
-    user: ${USER_NAME} ${PASS}
-    password: ${PASS}
-    """
-    r = rasa.utils.io.read_yaml(config_with_env_var)
-    assert r["user"] == "user pass" and r["password"] == "pass"
-
-
-def test_read_yaml_string_with_env_var_prefix():
-    config_with_env_var_prefix = """
-    user: db_${USER_NAME}
-    password: db_${PASS}
-    """
-    r = rasa.utils.io.read_yaml(config_with_env_var_prefix)
-    assert r["user"] == "db_user" and r["password"] == "db_pass"
-
-
-def test_read_yaml_string_with_env_var_postfix():
-    config_with_env_var_postfix = """
-    user: ${USER_NAME}_admin
-    password: ${PASS}_admin
-    """
-    r = rasa.utils.io.read_yaml(config_with_env_var_postfix)
-    assert r["user"] == "user_admin" and r["password"] == "pass_admin"
-
-
-def test_read_yaml_string_with_env_var_infix():
-    config_with_env_var_infix = """
-    user: db_${USER_NAME}_admin
-    password: db_${PASS}_admin
-    """
-    r = rasa.utils.io.read_yaml(config_with_env_var_infix)
-    assert r["user"] == "db_user_admin" and r["password"] == "db_pass_admin"
-
-
-def test_read_yaml_string_with_env_var_not_exist():
-    config_with_env_var_not_exist = """
-    user: ${USER_NAME}
-    password: ${PASSWORD}
-    """
-    with pytest.raises(ValueError):
-        rasa.utils.io.read_yaml(config_with_env_var_not_exist)
-
-
-@pytest.mark.parametrize("file, parents", [("A/test.md", "A"), ("A", "A")])
-def test_file_in_path(file, parents):
-    assert rasa.utils.io.is_subdirectory(file, parents)
-
-
-@pytest.mark.parametrize(
-    "file, parents", [("A", "A/B"), ("B", "A"), ("A/test.md", "A/B"), (None, "A")]
-)
-def test_file_not_in_path(file, parents):
-    assert not rasa.utils.io.is_subdirectory(file, parents)
