+++ tests/nlu/base/test_utils.py
@@ -4,7 +4,6 @@ import os
 import pickle
 import pytest
 import tempfile
-import rasa.utils.io
 from rasa.nlu import utils
 from rasa.nlu.utils import (
     create_dir,
@@ -16,8 +15,6 @@ from rasa.nlu.utils import (
     write_json_to_file,
     write_to_file,
 )
-from rasa.utils.endpoints import EndpointConfig
-import rasa.utils.io as io_utils
 
 
 @pytest.fixture(scope="function")
@@ -116,123 +113,3 @@ def test_remove_model_invalid(empty_model_dir):
 def test_is_url():
     assert not is_url("./some/file/path")
     assert is_url("https://rasa.com/")
-
-
-def test_environment_variable_not_existing():
-    content = "model: \n  test: ${variable}"
-    with pytest.raises(ValueError):
-        rasa.utils.io.read_yaml(content)
-
-
-def test_environment_variable_dict_without_prefix_and_postfix():
-    os.environ["variable"] = "test"
-    content = "model: \n  test: ${variable}"
-
-    result = rasa.utils.io.read_yaml(content)
-
-    assert result["model"]["test"] == "test"
-
-
-def test_environment_variable_in_list():
-    os.environ["variable"] = "test"
-    content = "model: \n  - value\n  - ${variable}"
-
-    result = rasa.utils.io.read_yaml(content)
-
-    assert result["model"][1] == "test"
-
-
-def test_environment_variable_dict_with_prefix():
-    os.environ["variable"] = "test"
-    content = "model: \n  test: dir/${variable}"
-
-    result = rasa.utils.io.read_yaml(content)
-
-    assert result["model"]["test"] == "dir/test"
-
-
-def test_environment_variable_dict_with_postfix():
-    os.environ["variable"] = "test"
-    content = "model: \n  test: ${variable}/dir"
-
-    result = rasa.utils.io.read_yaml(content)
-
-    assert result["model"]["test"] == "test/dir"
-
-
-def test_environment_variable_dict_with_prefix_and_with_postfix():
-    os.environ["variable"] = "test"
-    content = "model: \n  test: dir/${variable}/dir"
-
-    result = rasa.utils.io.read_yaml(content)
-
-    assert result["model"]["test"] == "dir/test/dir"
-
-
-def test_emojis_in_yaml():
-    test_data = """
-    data:
-        - one 😁💯 👩🏿‍💻👨🏿‍💻
-        - two £ (?u)\\b\\w+\\b f\u00fcr
-    """
-    actual = rasa.utils.io.read_yaml(test_data)
-
-    assert actual["data"][0] == "one 😁💯 👩🏿‍💻👨🏿‍💻"
-    assert actual["data"][1] == "two £ (?u)\\b\\w+\\b für"
-
-
-def test_emojis_in_tmp_file():
-    test_data = """
-        data:
-            - one 😁💯 👩🏿‍💻👨🏿‍💻
-            - two £ (?u)\\b\\w+\\b f\u00fcr
-        """
-    test_file = io_utils.create_temporary_file(test_data)
-    with io.open(test_file, mode="r", encoding="utf-8") as f:
-        content = f.read()
-    actual = rasa.utils.io.read_yaml(content)
-
-    assert actual["data"][0] == "one 😁💯 👩🏿‍💻👨🏿‍💻"
-    assert actual["data"][1] == "two £ (?u)\\b\\w+\\b für"
-
-
-def test_read_emojis_from_json():
-    import json
-
-    d = {"text": "hey 😁💯 👩🏿‍💻👨🏿‍💻🧜‍♂️(?u)\\b\\w+\\b} f\u00fcr"}
-    json_string = json.dumps(d, indent=2)
-
-    s = rasa.utils.io.read_yaml(json_string)
-
-    expected = "hey 😁💯 👩🏿‍💻👨🏿‍💻🧜‍♂️(?u)\\b\\w+\\b} für"
-    assert s.get("text") == expected
-
-
-def test_bool_str():
-    test_data = """
-    one: "yes"
-    two: "true"
-    three: "True"
-    """
-
-    actual = rasa.utils.io.read_yaml(test_data)
-
-    assert actual["one"] == "yes"
-    assert actual["two"] == "true"
-    assert actual["three"] == "True"
-
-
-def test_default_token_name():
-    test_data = {"url": "http://test", "token": "token"}
-
-    actual = EndpointConfig.from_dict(test_data)
-
-    assert actual.token_name == "token"
-
-
-def test_custom_token_name():
-    test_data = {"url": "http://test", "token": "token", "token_name": "test_token"}
-
-    actual = EndpointConfig.from_dict(test_data)
-
-    assert actual.token_name == "test_token"
